Update Main.java
Added print statements to print the correct message as stated on the PDF.
I also removed the pre-set words we originally had to test getWordLadderBFS and switched it out with the user input to test out other words and if it could work on with different length words. Looks like it works, probably need to write some junits later.